requirements.txt
ultralytics>=8.3.0
pandas>=2.0.0
pyyaml>=6.0.0
opencv-python>=4.8.0
pillow>=10.0.0
tqdm>=4.66.0
